Commit 2
Task 2 complete. JUnit test returned green after changing the line:

this.x = y;

to

this.x = x;

In the Point class. Methods getRectangle() and testGetRectangle() were added to the Rectangle and TestRectangle class respectively. This makes it easier to find the previous error because it lets the user know somethind is specifically wrong with the rectangle points.